Issue #231: Move endpoint related to payments from communities to payments
