+++ api/src/main/resources/public/swagger/api.yaml
@@ -956,7 +956,7 @@ paths:
       security:
         - bearerAuth: [ ]
       tags:
-        - Communities
+        - Payments
       description: Get all payments scheduled by the specified admin
       operationId: listAllAdminScheduledPayments
       parameters:
