+++ service/src/test/java/com/myhome/controllers/PaymentControllerTest.java
@@ -27,16 +27,21 @@ import com.myhome.domain.HouseMember;
 import com.myhome.domain.HouseMemberDocument;
 import com.myhome.domain.Payment;
 import com.myhome.domain.User;
+import com.myhome.model.AdminPayment;
 import com.myhome.model.HouseMemberDto;
+import com.myhome.model.ListAdminPaymentsResponse;
 import com.myhome.model.ListMemberPaymentsResponse;
 import com.myhome.model.MemberPayment;
 import com.myhome.services.CommunityService;
 import com.myhome.services.PaymentService;
+import com.myhome.utils.PageInfo;
 import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
-import java.util.Arrays;
+import java.util.ArrayList;
+import java.util.Collections;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Optional;
 import java.util.Set;
 import org.junit.jupiter.api.BeforeEach;
@@ -44,6 +49,9 @@ import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
+import org.springframework.data.domain.PageImpl;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 
@@ -75,7 +83,9 @@ class PaymentControllerTest {
   private static final String COMMUNITY_HOUSE_ID = "5";
   private static final String TEST_MEMBER_ID = "2";
   private static final String TEST_ID = "3";
-  private static final String TEST_COMMUNITY_ID = "4";
+  private static final String TEST_TEST_ID = "4";
+
+  private static final Pageable TEST_PAGEABLE = PageRequest.of(1, 10);
 
   @Mock
   private PaymentService paymentService;
@@ -83,11 +93,11 @@ class PaymentControllerTest {
   @Mock
   private SchedulePaymentApiMapper paymentApiMapper;
 
-  @InjectMocks
-  private PaymentController paymentController;
-
   @Mock
   private CommunityService communityService;
+
+  @InjectMocks
+  private PaymentController paymentController;
 
   @BeforeEach
   private void init() {
@@ -97,7 +107,7 @@ class PaymentControllerTest {
   private PaymentDto createTestPaymentDto() {
     UserDto userDto = UserDto.builder()
         .userId(TEST_ADMIN_ID)
-        .communityIds(new HashSet<>(Arrays.asList(TEST_COMMUNITY_ID)))
+        .communityIds(new HashSet<>(Collections.singletonList(TEST_TEST_ID)))
         .id(Long.valueOf(TEST_ADMIN_ID))
         .encryptedPassword(TEST_ADMIN_PASSWORD)
         .name(TEST_ADMIN_NAME)
@@ -124,13 +134,13 @@ class PaymentControllerTest {
     CommunityDto communityDto = new CommunityDto();
     communityDto.setName(TEST_COMMUNITY_NAME);
     communityDto.setDistrict(TEST_COMMUNITY_DISTRICT);
-    communityDto.setCommunityId(TEST_COMMUNITY_ID);
+    communityDto.setCommunityId(TEST_TEST_ID);
     return communityDto;
   }
 
   private Community getMockCommunity(Set<User> admins) {
     Community community =
-        new Community(admins, new HashSet<>(), TEST_COMMUNITY_NAME, TEST_COMMUNITY_ID,
+        new Community(admins, new HashSet<>(), TEST_COMMUNITY_NAME, TEST_TEST_ID,
             TEST_COMMUNITY_DISTRICT, new HashSet<>());
     User admin = new User(COMMUNITY_ADMIN_NAME, TEST_ADMIN_ID, COMMUNITY_ADMIN_EMAIL, false,
         COMMUNITY_ADMIN_PASSWORD, new HashSet<>(), new HashSet<>());
@@ -181,9 +191,10 @@ class PaymentControllerTest {
 
     EnrichedSchedulePaymentRequest enrichedRequest =
         new EnrichedSchedulePaymentRequest(TEST_TYPE, TEST_DESCRIPTION, TEST_RECURRING, TEST_CHARGE,
-            TEST_DUE_DATE, TEST_ADMIN_ID, Long.valueOf(1), TEST_ADMIN_NAME, TEST_ADMIN_EMAIL,
-            TEST_ADMIN_PASSWORD, new HashSet<>(Arrays.asList(TEST_COMMUNITY_ID)), TEST_MEMBER_ID,
-            Long.valueOf(2), "", TEST_MEMBER_NAME, COMMUNITY_HOUSE_ID);
+            TEST_DUE_DATE, TEST_ADMIN_ID, 1L, TEST_ADMIN_NAME, TEST_ADMIN_EMAIL,
+            TEST_ADMIN_PASSWORD, new HashSet<>(Collections.singletonList(TEST_TEST_ID)),
+            TEST_MEMBER_ID,
+            2L, "", TEST_MEMBER_NAME, COMMUNITY_HOUSE_ID);
     PaymentDto paymentDto = createTestPaymentDto();
     com.myhome.model.SchedulePaymentResponse response =
         new com.myhome.model.SchedulePaymentResponse()
@@ -421,7 +432,8 @@ class PaymentControllerTest {
         .willReturn(Optional.empty());
 
     //when
-    ResponseEntity<ListMemberPaymentsResponse> responseEntity = paymentController.listAllMemberPayments(TEST_MEMBER_ID);
+    ResponseEntity<ListMemberPaymentsResponse> responseEntity =
+        paymentController.listAllMemberPayments(TEST_MEMBER_ID);
 
     //then
     assertEquals(HttpStatus.NOT_FOUND, responseEntity.getStatusCode());
@@ -456,14 +468,15 @@ class PaymentControllerTest {
             .charge(TEST_CHARGE)
             .dueDate(TEST_DUE_DATE));
 
-
-    ListMemberPaymentsResponse expectedResponse = new ListMemberPaymentsResponse().payments(paymentResponses);
+    ListMemberPaymentsResponse expectedResponse =
+        new ListMemberPaymentsResponse().payments(paymentResponses);
 
     given(paymentApiMapper.memberPaymentSetToRestApiResponseMemberPaymentSet(payments))
         .willReturn(paymentResponses);
 
     // when
-    ResponseEntity<ListMemberPaymentsResponse> responseEntity = paymentController.listAllMemberPayments(TEST_MEMBER_ID);
+    ResponseEntity<ListMemberPaymentsResponse> responseEntity =
+        paymentController.listAllMemberPayments(TEST_MEMBER_ID);
 
     // then
     assertEquals(HttpStatus.OK, responseEntity.getStatusCode());
@@ -472,4 +485,114 @@ class PaymentControllerTest {
     verify(paymentApiMapper).memberPaymentSetToRestApiResponseMemberPaymentSet(payments);
   }
 
+  @Test
+  void shouldGetAdminPaymentsSuccess() {
+    // given
+    com.myhome.model.SchedulePaymentRequest request =
+        new com.myhome.model.SchedulePaymentRequest()
+            .type(TEST_TYPE)
+            .description(TEST_DESCRIPTION)
+            .recurring(TEST_RECURRING)
+            .charge(TEST_CHARGE)
+            .dueDate(TEST_DUE_DATE)
+            .adminId(TEST_ADMIN_ID)
+            .memberId(TEST_MEMBER_ID);
+    PaymentDto paymentDto = createTestPaymentDto();
+
+    given(paymentService.schedulePayment(paymentDto))
+        .willReturn(paymentDto);
+
+    List<Payment> payments = new ArrayList<>();
+    Payment mockPayment = getMockPayment();
+    payments.add(mockPayment);
+
+    Set<String> adminIds = new HashSet<>();
+    adminIds.add(TEST_ADMIN_ID);
+
+    Set<User> admins = new HashSet<>();
+
+    Community community = getMockCommunity(admins);
+
+    CommunityDto communityDto = createTestCommunityDto();
+
+    given(communityService.createCommunity(communityDto))
+        .willReturn(community);
+    given(communityService.getCommunityDetailsByIdWithAdmins(TEST_ID))
+        .willReturn(Optional.of(community));
+    given(paymentService.getPaymentsByAdmin(TEST_ADMIN_ID, TEST_PAGEABLE))
+        .willReturn(new PageImpl<>(payments));
+    given(communityService.addAdminsToCommunity(TEST_ID, adminIds))
+        .willReturn(Optional.of(community));
+
+    Set<AdminPayment> responsePayments = new HashSet<>();
+    responsePayments.add(
+        new AdminPayment().adminId(TEST_ADMIN_ID)
+            .paymentId(TEST_ID)
+            .charge(TEST_CHARGE)
+            .dueDate(TEST_DUE_DATE)
+    );
+
+    ListAdminPaymentsResponse expectedResponse =
+        new ListAdminPaymentsResponse()
+            .payments(responsePayments)
+            .pageInfo(PageInfo.of(TEST_PAGEABLE, new PageImpl<>(payments)));
+
+    given(paymentApiMapper.adminPaymentSetToRestApiResponseAdminPaymentSet(new HashSet<>(payments)))
+        .willReturn(responsePayments);
+
+    //when
+    ResponseEntity<ListAdminPaymentsResponse> responseEntity =
+        paymentController.listAllAdminScheduledPayments(TEST_ID, TEST_ADMIN_ID,
+            TEST_PAGEABLE);
+
+    //then
+    assertEquals(HttpStatus.OK, responseEntity.getStatusCode());
+    assertEquals(expectedResponse, responseEntity.getBody());
+    verify(communityService).getCommunityDetailsByIdWithAdmins(TEST_ID);
+    verify(paymentService).getPaymentsByAdmin(TEST_ADMIN_ID, TEST_PAGEABLE);
+    verify(paymentApiMapper).adminPaymentSetToRestApiResponseAdminPaymentSet(
+        new HashSet<>(payments));
+  }
+
+  @Test
+  void shouldReturnNotFoundWhenAdminIsNotInCommunity() {
+    //given
+    final String notAdminFromCommunity = "2";
+    Community community = getMockCommunity(new HashSet<>());
+    given(communityService.getCommunityDetailsByIdWithAdmins(TEST_ID))
+        .willReturn(Optional.of(community));
+
+    //when
+    ResponseEntity<ListAdminPaymentsResponse> responseEntity =
+        paymentController.listAllAdminScheduledPayments(TEST_ID, notAdminFromCommunity,
+            TEST_PAGEABLE);
+
+    //then
+    assertEquals(HttpStatus.NOT_FOUND, responseEntity.getStatusCode());
+    assertNull(responseEntity.getBody());
+    verify(communityService).getCommunityDetailsByIdWithAdmins(TEST_ID);
+    verifyNoInteractions(paymentService);
+  }
+
+  @Test
+  void shouldThrowExceptionWhenCommunityNotExists() {
+    //given
+    String expectedExceptionMessage = "Community with given id not exists: " + TEST_ID;
+
+    given(communityService.getCommunityDetailsByIdWithAdmins(TEST_ID))
+        .willReturn(Optional.empty());
+
+    //when
+    final RuntimeException runtimeException = assertThrows(
+        RuntimeException.class,
+        () -> paymentController.listAllAdminScheduledPayments(TEST_ID, TEST_ADMIN_ID,
+            TEST_PAGEABLE)
+    );
+
+    //then
+    assertEquals(expectedExceptionMessage, runtimeException.getMessage());
+    verify(communityService).getCommunityDetailsByIdWithAdmins(TEST_ID);
+    verifyNoInteractions(paymentService);
+    verifyNoInteractions(paymentApiMapper);
+  }
 }
